Optional slurm business

+++ 01_process_mms.py
@@ -55,6 +55,7 @@ for item in ['L1','L2','L3']:
         polyorder = config[item]['polyorder']
 
 
+slurm = config['SLURM']
 containers = config['CONTAINERS']
 tricolour_container = config[containers]['TRICOLOUR']
 wsclean_container = config[containers]['WSCLEAN']
@@ -129,9 +130,12 @@ for submsname in mslist:
     flag_name = 'TRICO1'+label
     flag_runfile = scripts_dir+'slurm_'+flag_name+'.sh'
     flag_logfile = flag_runfile.replace('.sh','.log').replace(scripts_dir,logs_dir)
-    flag_syscall = gen.tricolour(tricolour_container,submsname,tricolour_config,'DATA',residuals=False)
+    flag_syscall = gen.tricolour(slurm,tricolour_container,submsname,tricolour_config,'DATA',residuals=False)
     gen.write_slurm(flag_runfile,flag_logfile,flag_name,'03:00:00',16,'115GB',flag_syscall)
-    flag_run_command = flag_name+"=`sbatch "+flag_runfile+" | awk '{print $4}'`\n"
+    if slurm:
+        flag_run_command = flag_name+"=`sbatch "+flag_runfile+" | awk '{print $4}'`\n"
+    else:
+        flag_run_command = f'source {flag_runfile}'
     f.write(flag_run_command)
     master_job_list.append(flag_name)
     post_loop_dependencies.append(flag_name)
@@ -150,9 +154,12 @@ f.write('# Initial imaging with mask to get continuum model\n')
 wsclean_name = 'WSCLN1'
 wsclean_runfile = scripts_dir+'slurm_'+wsclean_name+'.sh'
 wsclean_logfile = wsclean_runfile.replace('.sh','.log').replace(scripts_dir,logs_dir)
-wsclean_syscall = gen.wsclean(wsclean_container,myms,'DATA',image_name,fitsmask,chansout,polyorder,dirty=False)
+wsclean_syscall = gen.wsclean(slurm,wsclean_container,myms,'DATA',image_name,fitsmask,chansout,polyorder,dirty=False)
 gen.write_slurm(wsclean_runfile,wsclean_logfile,wsclean_name,'24:00:00',32,'230GB',wsclean_syscall)
-run_command = wsclean_name+"=`sbatch -d afterok:"+post_loop_dependencies+" "+wsclean_runfile+" | awk '{print $4}'`\n"
+if slurm:
+    run_command = wsclean_name+"=`sbatch -d afterok:"+post_loop_dependencies+" "+wsclean_runfile+" | awk '{print $4}'`\n"
+else:
+    run_command = f'source {wsclean_runfile}'
 f.write(run_command)
 master_job_list.append(wsclean_name)
 
@@ -166,11 +173,14 @@ f.write('# Frequency interpolation of continuum model\n')
 smops_name = 'SMOPS'
 smops_runfile = scripts_dir+'slurm_'+smops_name+'.sh'
 smops_logfile = smops_runfile.replace('.sh','.log').replace(scripts_dir,logs_dir)
-smops_syscall = gen.smops(smops_container,myms,image_name,smopschans,3,smopsmodel)+'\n'
+smops_syscall = gen.smops(slurm,smops_container,myms,image_name,smopschans,3,smopsmodel)+'\n'
 for i in range(0,nsubms): # symlink the interpolated models into the temp folders
     smops_syscall += 'ln -s '+smopsmodel+'* temp_'+str(i).zfill(3)+'/\n'
 gen.write_slurm(smops_runfile,smops_logfile,smops_name,'03:00:00',16,'115GB',smops_syscall)
-run_command = smops_name+"=`sbatch -d afterok:$"+wsclean_name+" "+smops_runfile+" | awk '{print $4}'`\n"
+if slurm:
+    run_command = smops_name+"=`sbatch -d afterok:$"+wsclean_name+" "+smops_runfile+" | awk '{print $4}'`\n"
+else:
+    run_command = f'source {smops_runfile}'
 f.write(run_command)
 master_job_list.append(smops_name)
 
@@ -196,9 +206,12 @@ for submsname in mslist:
     predict_name = 'PRDICT'+label
     predict_runfile = scripts_dir+'slurm_'+predict_name+'.sh'
     predict_logfile = predict_runfile.replace('.sh','.log').replace(scripts_dir,logs_dir)
-    predict_syscall = gen.predict(wsclean_container,submsname,tempdir,smopschans,smopsmodel.split('/')[-1])
+    predict_syscall = gen.predict(slurm,wsclean_container,submsname,tempdir,smopschans,smopsmodel.split('/')[-1])
     gen.write_slurm(predict_runfile,predict_logfile,predict_name,'12:00:00',16,'115GB',predict_syscall)
-    predict_run_command = predict_name+"=`sbatch -d afterok:$"+smops_name+" "+predict_runfile+" | awk '{print $4}'`\n"
+    if slurm:
+        predict_run_command = predict_name+"=`sbatch -d afterok:$"+smops_name+" "+predict_runfile+" | awk '{print $4}'`\n"
+    else:
+        predict_run_command = f'source {predict_runfile}'
     f.write(predict_run_command)
     master_job_list.append(predict_name)
 
@@ -211,9 +224,12 @@ for submsname in mslist:
     selfcal_name = 'CUBICL'+label
     selfcal_runfile = scripts_dir+'slurm_'+selfcal_name+'.sh'
     selfcal_logfile = selfcal_runfile.replace('.sh','.log').replace(scripts_dir,logs_dir)
-    selfcal_syscall = gen.cubical(cubical_container,submsname,cubical_config,nchan,timechunk)
+    selfcal_syscall = gen.cubical(slurm,cubical_container,submsname,cubical_config,nchan,timechunk)
     gen.write_slurm(selfcal_runfile,selfcal_logfile,selfcal_name,'08:00:00',32,'230GB',selfcal_syscall)
-    selfcal_run_command = selfcal_name+"=`sbatch -d afterok:$"+predict_name+" "+selfcal_runfile+" | awk '{print $4}'`\n"
+    if slurm:
+        selfcal_run_command = selfcal_name+"=`sbatch -d afterok:$"+predict_name+" "+selfcal_runfile+" | awk '{print $4}'`\n"
+    else:
+        selfcal_run_command = f'source {selfcal_runfile}'
     f.write(selfcal_run_command)
     master_job_list.append(selfcal_name)
 
@@ -226,9 +242,12 @@ for submsname in mslist:
     flag_name = 'TRICO2'+label
     flag_runfile = scripts_dir+'slurm_'+flag_name+'.sh'
     flag_logfile = flag_runfile.replace('.sh','.log').replace(scripts_dir,logs_dir)
-    flag_syscall = gen.tricolour(tricolour_container,submsname,tricolour_config,'CORRECTED_DATA',residuals=True)
+    flag_syscall = gen.tricolour(slurm,tricolour_container,submsname,tricolour_config,'CORRECTED_DATA',residuals=True)
     gen.write_slurm(flag_runfile,flag_logfile,flag_name,'03:00:00',16,'115GB',flag_syscall)
-    flag_run_command = flag_name+"=`sbatch -d afterok:$"+selfcal_name+" "+flag_runfile+" | awk '{print $4}'`\n"
+    if slurm:
+        flag_run_command = flag_name+"=`sbatch -d afterok:$"+selfcal_name+" "+flag_runfile+" | awk '{print $4}'`\n"
+    else:
+        flag_run_command = f'source {flag_runfile}'
     f.write(flag_run_command)
     master_job_list.append(flag_name)
     post_loop_dependencies.append(flag_name)
@@ -249,24 +268,28 @@ f.write('# Image residuals to check continuum subtraction and selfcal\n')
 wsclean_name = 'WSCLN2'
 wsclean_runfile = scripts_dir+'slurm_'+wsclean_name+'.sh'
 wsclean_logfile = wsclean_runfile.replace('.sh','.log').replace(scripts_dir,logs_dir)
-wsclean_syscall = gen.wsclean(wsclean_container,myms,'CORRECTED_DATA',resid_name,fitsmask,chansout,polyorder,dirty=True)
+wsclean_syscall = gen.wsclean(slurm,wsclean_container,myms,'CORRECTED_DATA',resid_name,fitsmask,chansout,polyorder,dirty=True)
 gen.write_slurm(wsclean_runfile,wsclean_logfile,wsclean_name,'24:00:00',32,'230GB',wsclean_syscall)
-run_command = wsclean_name+"=`sbatch -d afterok:"+post_loop_dependencies+" "+wsclean_runfile+" | awk '{print $4}'`\n"
-f.write(run_command)
+if slurm:
+    wsclean_run_command = wsclean_name+"=`sbatch -d afterok:"+post_loop_dependencies+" "+wsclean_runfile+" | awk '{print $4}'`\n"
+else:
+    wsclean_run_command = f'source {wsclean_runfile}'
+f.write(wsclean_run_command)
 master_job_list.append(wsclean_name)
 
 
 # ------------------------------------------------------------------
 # Write the kill commands and close the submit file
 
-f.write('\n# --------------------------------------------------\n')
-f.write('# Kill commands\n')
-for job_id in master_job_list:
-    if job_id == master_job_list[0]:
-        kill = 'echo "scancel $'+job_id+' #'+job_id+'" > '+kill_file+'\n'
-    else:
-        kill = 'echo "scancel $'+job_id+' #'+job_id+'" >> '+kill_file+'\n'
-    f.write(kill)
-f.close()
+if slurm:
+    f.write('\n# --------------------------------------------------\n')
+    f.write('# Kill commands\n')
+    for job_id in master_job_list:
+        if job_id == master_job_list[0]:
+            kill = 'echo "scancel $'+job_id+' #'+job_id+'" > '+kill_file+'\n'
+        else:
+            kill = 'echo "scancel $'+job_id+' #'+job_id+'" >> '+kill_file+'\n'
+        f.write(kill)
+    f.close()
 print('Submit jobs                 : source %s' %submit_file)
 print('----------------------------+---------------------------------------\n')
